Added rocs::core::Configuration::getChildren.
getChildren returns the path to all childs of the given argument.
Had to implement tree-recursion code since the previous getChilds
was not usable (it skipped one node name).

Anyway the older function was a bit useless and theres plenty of
code on it using: vector<ptree> meaning we are copying the tree a
few times each time we list elements on it.

This class needs serious restructure for loading huge ammounts of data.