Rename Factor to FactorData.
The targetd class pretends to record all data related to a factor that might be reused across a graph.
The name class Factor will be used to represent a given instance of that one.
This is similar to VariableType being reused across diferent variable instances and keeps name scheme easy.